check fieldsArray instead of options.fields
I had strange issues while checking options.fields. It looked like
options.fields got changed after the cursor was initialized.
options.fields was an empty object but the fieldsArray was undefined.
Since there is only a check if options.fields is defined, the result
was an exception, that fieldsArray is undefined.